refactor: fix drill parse and improve test render page
