+++ packages/parser/package.json
@@ -5,7 +5,7 @@
   },
   "version": "5.0.0-next.0",
   "description": "Gerber and NC drill file parser",
-  "main": "./dist/tracespace-parser.umd.js",
+  "main": "./dist/tracespace-parser.umd.cjs",
   "module": "./dist/tracespace-parser.es.js",
   "types": "./lib/index.d.ts",
   "exports": {
@@ -13,9 +13,10 @@
       "types": "./lib/index.d.ts",
       "source": "./src/index.ts",
       "import": "./dist/tracespace-parser.es.js",
-      "require": "./dist/tracespace-parser.umd.js"
+      "require": "./dist/tracespace-parser.umd.cjs"
     }
   },
+  "type": "module",
   "sideEffects": false,
   "repository": {
     "type": "git",
