+++ packages/parser/src/syntax/__tests__/drill.test.ts
@@ -8,7 +8,7 @@ import {
   position as pos,
   simplifyTokens,
 } from '../../__tests__/helpers'
-import {matchSyntax, MatchState} from '..'
+import {matchSyntax, grammar, MatchState} from '..'
 
 import {
   DRILL,
@@ -599,7 +599,7 @@ describe('drill syntax matches', () => {
       lexer.reset(source)
 
       for (const token of lexer) {
-        result = matchSyntax(result, token)
+        result = matchSyntax(result, token, grammar)
       }
 
       const {tokens, nodes, filetype} = result!
