+++ packages/parser/vite.config.ts
@@ -1,6 +1,6 @@
 import {defineConfig} from 'vite'
 
-import {baseConfig} from '../../config/vite.config.base'
+import {baseConfig, libraryFilename} from '../../config/vite.config.base'
 
 export default defineConfig({
   ...baseConfig,
@@ -8,7 +8,7 @@ export default defineConfig({
     lib: {
       entry: 'src/index.ts',
       name: 'TracespaceParser',
-      fileName: 'tracespace-parser',
+      fileName: libraryFilename('tracespace-parser'),
     },
   },
 })
