+++ packages/whats-that-gerber/package.json
@@ -2,7 +2,7 @@
   "name": "whats-that-gerber",
   "version": "5.0.0-next.0",
   "description": "Identify Gerber and drill files by filename",
-  "main": "./dist/whats-that-gerber.umd.js",
+  "main": "./dist/whats-that-gerber.umd.cjs",
   "module": "./dist/whats-that-gerber.es.js",
   "types": "./lib/index.d.ts",
   "exports": {
@@ -10,17 +10,11 @@
       "types": "./lib/index.d.ts",
       "source": "./src/index.ts",
       "import": "./dist/whats-that-gerber.es.js",
-      "require": "./dist/whats-that-gerber.umd.js"
+      "require": "./dist/whats-that-gerber.umd.cjs"
     }
   },
+  "type": "module",
   "sideEffects": false,
-  "files": [
-    "src",
-    "lib",
-    "esm",
-    "cjs",
-    "umd"
-  ],
   "scripts": {
     "build": "vite build",
     "clean": "rimraf dist"
