+++ packages/whats-that-gerber/vite.config.ts
@@ -1,11 +1,14 @@
 import {defineConfig} from 'vite'
 
+import {baseConfig, libraryFilename} from '../../config/vite.config.base'
+
 export default defineConfig({
+  ...baseConfig,
   build: {
     lib: {
       entry: 'src/index.ts',
       name: 'WhatsThatGerber',
-      fileName: (format: string) => `whats-that-gerber.${format}.js`,
+      fileName: libraryFilename('whats-that-gerber'),
     },
   },
 })
