+++ packages/xml-id/vite.config.ts
@@ -1,11 +1,14 @@
 import {defineConfig} from 'vite'
 
+import {baseConfig, libraryFilename} from '../../config/vite.config.base'
+
 export default defineConfig({
+  ...baseConfig,
   build: {
     lib: {
       entry: 'src/index.ts',
       name: 'TracespaceXmlId',
-      fileName: (format: string) => `tracespace-xml-id.${format}.js`,
+      fileName: libraryFilename('tracespace-xml-id'),
     },
   },
 })
