+++ www/src/base/_default.page.server.tsx
@@ -10,10 +10,9 @@ export function render(pageContext: PageContext): unknown {
   const documentHtml = escapeInject`<!DOCTYPE html>
     <html lang="en">
       <head>
-        <meta charset="UTF-8" />
+        <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta name="description" content="Hello description" />
-        <title>Hello title</title>
+        <title>tracespace</title>
       </head>
       <body>
         ${dangerouslySkipEscape(pageHtml)}
